Moved validation files to utils
