+++ Common/tests/test_forms.py
@@ -34,21 +34,3 @@ class ProductImageFormTests(TestCase):
 
         form = ProductImageForm(data=data)
         self.assertTrue(form.is_valid())
-
-    # def test_image_upload_form_missing_image(self):
-    #     """ Test image upload form when file is missing """
-
-    #     data = {
-    #         "alt_text": "Test Image",
-    #     }
-
-    #     form = ProductImageForm(data=data)
-
-    #     self.assertFalse(form.is_valid())
-    #     self.assertIn("file", form.errors)
-
-
-class ImageUploadFormTests(TestCase):
-    """ Test the ImageUploadForm """
-
-
